typing + doc + (close #27)

+++ knowledge_clustering/file_updater.py
@@ -6,20 +6,23 @@
 # the user has to manually merge
 # and a prompt is offered using click.
 #
+
+from __future__ import annotations
+
 import hashlib
 import tempfile
 import os
 import click
 
 
-def hash_file(filepath):
+def hash_file(filepath: str):
     """
     Compute a hash of the content of the
     given filepath
     """
     with open(filepath, "rb") as f:
         file_hash = hashlib.blake2b()
-        chunk = f.read(8192)
+        chunk: bytes = f.read(8192)
         while chunk:
             file_hash.update(chunk)
             chunk = f.read(8192)
@@ -34,12 +37,12 @@ class AtomicUpdate:
     """
 
     def __init__(self, filename, original_hash=None):
-        self.filename = filename
+        self.filename: str = filename
         self.hash = hash_file(filename)
         self.ctx = tempfile.NamedTemporaryFile(mode="w", dir=os.getcwd(), delete=False)
         self.tmp = None
         if (
-            original_hash.hexdigest() != None
+            original_hash is not None
             and original_hash.hexdigest() != self.hash.hexdigest()
         ):
             click.confirm(
@@ -57,20 +60,21 @@ class AtomicUpdate:
 
     def __exit__(self, type, value, traceback):
         nh = hash_file(self.filename)
-        if nh.hexdigest() != self.hash.hexdigest():
-            print(f"{nh.hexdigest()} ≠ {self.hash.hexdigest()}")
-            b = click.confirm(
-                f"File {self.filename} has been modified\
-                during the run of \
-                the program, erase anyway?",
-                default=None,
-                abort=False,
-                prompt_suffix=": ",
-                show_default=True,
-                err=False,
-            )
-            if b is False:
-                print(f"Temporary file accessible at {self.tmp.name}")
-                return self.ctx.__exit__(type, value, traceback)
-        os.replace(self.tmp.name, self.filename)
+        if self.tmp is not None:
+            if nh.hexdigest() != self.hash.hexdigest():
+                print(f"{nh.hexdigest()} ≠ {self.hash.hexdigest()}")
+                b = click.confirm(
+                    f"File {self.filename} has been modified\
+                    during the run of \
+                    the program, erase anyway?",
+                    default=None,
+                    abort=False,
+                    prompt_suffix=": ",
+                    show_default=True,
+                    err=False,
+                )
+                if b is False:
+                    print(f"Temporary file accessible at {self.tmp.name}")
+                    return self.ctx.__exit__(type, value, traceback)
+            os.replace(self.tmp.name, self.filename)
         return self.ctx.__exit__(type, value, traceback)
